Completed CRUD for companies

+++ routes/companies.router.js
@@ -24,11 +24,11 @@ router.get('/',password.authenticate('jwt', {session: false}),
     }
 });
 
-/*
+
 router.post('/',
 password.authenticate('jwt', {session: false}),
-checkRoles('admin'),
-  validationHandler(createCustomerSchema, 'body'),
+checkRoles(1),
+  validationHandler(createCompanySchema, 'body'),
   async (req, res, next) => {
     try {
       const body = req.body;
@@ -39,9 +39,10 @@ checkRoles('admin'),
   }
 );
 
+
 router.patch('/:id',
-  validationHandler(getCustomerSchema, 'params'),
-  validationHandler(updateCustomerSchema, 'body'),
+  validationHandler(getCompanySchema, 'params'),
+  validationHandler(updateCompanySchema, 'body'),
   async (req, res, next) => {
     try {
       const { id } = req.params;
@@ -54,16 +55,18 @@ router.patch('/:id',
 );
 
 router.delete('/:id',
-  validationHandler(getCustomerSchema, 'params'),
+  validationHandler(getCompanySchema, 'params'),
   async (req, res, next) => {
     try {
       const { id } = req.params;
+      if (id == 1) {
+        throw Error("Can't delete main company");
+      }
       res.status(200).json(await service.delete(id));
     } catch (error) {
       next(error);
     }
   }
 );
-*/
 
 module.exports = router;
